file rename and minor refactor to comply with raii
