+++ src/render/Swapchain.hpp
@@ -20,6 +20,7 @@ class Swapchain {
     private:
         SwapChainSupportDetails findSwapChainSupportDetails();
     public:
+        Swapchain();
         inline static VkSwapchainKHR swapchain;
         inline static VkExtent2D swapchainExtent;
         // needs refactor into Image class
