+++ src/render/main.cpp
@@ -1,4 +1,4 @@
-#include "app.hpp"
+#include "App.hpp"
 #include <iostream>
 
 int main()
